Adding FlySystem as fileSystem abstraction + Monolog + Using Redis as caching file system

+++ src/Core/Service/ImageResizer.php
@@ -2,13 +2,17 @@
 
 namespace Core\Service;
 
+use League\Flysystem\Filesystem;
+use pastuhov\Command\Command;
+use Monolog\Logger;
+
 
 class ImageResizer
 {
     /**
-     * @var string
+     * @var Filesystem
      */
-    protected $rootDir;
+    protected $filesystem;
 
     /**
      * @var string
@@ -16,13 +20,21 @@ class ImageResizer
     protected $mozJpegExecutable;
 
     /**
+     * @var Logger
+     */
+    protected $logger;
+
+    /**
      * Resizer constructor.
      * @param array $params
+     * @param Filesystem $filesystem
+     * @param Logger $logger
      */
-    public function __construct($params)
+    public function __construct($params, Filesystem $filesystem, Logger $logger)
     {
         $this->mozJpegExecutable = $params['mozjpeg_path'];
-        $this->rootDir = $params['root_dir'];
+        $this->filesystem = $filesystem;
+        $this->logger = $logger;
     }
 
     /**
@@ -32,27 +44,69 @@ class ImageResizer
      */
     public function resize($sourceFile, $options = [])
     {
-        $newFileName = null;
-        try {
-            $tmpFile = $this->rootDir . '/var/tmp/' . uniqid("", true);
-            file_put_contents($tmpFile, file_get_contents($sourceFile));
-            $size = $options['width'] . 'x' . $options['height'];
-            $unsharp = $options['unsharp'];
-            $quality = $options['quality'];
+        $size = $options['width'] . 'x' . $options['height'];
+        $unsharp = $options['unsharp'];
+        $quality = $options['quality'];
+        $newFileName = md5($sourceFile . $size . $unsharp . $quality);
 
-            $newFileName = $this->rootDir . '/var/tmp/' . time() . '-' . uniqid("", true) . '.jpeg';
-            $command = "/usr/bin/nice /usr/bin/convert {$tmpFile}'[{$size}]' -strip -limit thread 1 -gravity center -extent {$size} -sampling-factor 1x1 -unsharp {$unsharp} -filter Lanczos ";
+        if (!$this->filesystem->has($newFileName)) {
+            $newFilePath = TMP_DIR . $newFileName;
+
+            $tmpFile = $this->saveTmpFile($sourceFile);
+
+            $command = "/usr/bin/convert {tmpFile}'[{size}]' \\
+            -strip -limit thread 1 -gravity center -extent {size} \\
+            -sampling-factor 1x1 \\
+            -unsharp {unsharp} -filter Lanczos ";
+            $params = [
+                'tmpFile' => $tmpFile,
+                'size' => $size,
+                'unsharp' => $unsharp,
+            ];
 
             if (is_executable($this->mozJpegExecutable)) {
-                $command .= "TGA:- | {$this->mozJpegExecutable} -quality {$quality} -outfile ${newFileName} -targa";
+                $command .= "TGA:- | {mozJpegExecutable} -quality {quality} -outfile {newFilePath} -targa";
+                $params['mozJpegExecutable'] = $this->mozJpegExecutable;
             } else {
-                $command .= "-quality {$quality} {$newFileName}";
+                $command .= "-quality {quality} {newFilePath}";
             }
-            exec($command);
-            unlink($tmpFile);
+            $params['quality'] = $quality;
+            $params['newFilePath'] = $newFilePath;
+            try {
+                Command::exec(
+                    $command,
+                    $params
+                );
 
-        } catch (\Exception $e) {
+                $this->filesystem->write($newFileName, stream_get_contents(fopen($newFilePath, 'r')));
+                unlink($tmpFile);
+                unlink($newFilePath);
+
+            } catch (\Exception $e) {
+                var_dump('Exception: ' . $e->getMessage());
+                exit;
+            }
         }
-        return $newFileName;
+        return $this->filesystem->read($newFileName);
+    }
+
+    /**
+     * Save given image in tmp file and return the path
+     * @param $fileUrl
+     * @return string
+     * @throws \Exception
+     */
+    public function saveTmpFile($fileUrl)
+    {
+        if (!$resource = @fopen($fileUrl, "r")) {
+            throw  new \Exception('Error occured while trying to read the file Url');
+        }
+        $content = "";
+        while ($line = fread($resource, 1024)) {
+            $content .= $line;
+        }
+        $tmpFile = TMP_DIR . uniqid("", true);
+        file_put_contents($tmpFile, $content);
+        return $tmpFile;
     }
 }
